+++ Dockerfile
@@ -1,8 +1,14 @@
 FROM golang:1.13-alpine
-RUN mkdir /app
-WORKDIR /app
-COPY ./* ./
-RUN rm -rf testdata && rm *_test.go && go build -o raindrops && mv raindrops /bin/raindrops
+RUN apk add git
+RUN mkdir /raindrops
+WORKDIR /raindrops
+COPY go.mod .
+COPY go.sum .
+ENV GO111MODULE=on
+RUN go mod download
+COPY . ./
+RUN rm -rf ./testdata &&  find ./ |grep ".*_test.go" |xargs rm
+RUN go build -o raindrops ./cmd  && mv raindrops /bin/raindrops
 RUN chmod +x docker-entrypoint.sh && mv docker-entrypoint.sh /bin/ && rm -rf ./*
 ENTRYPOINT ["docker-entrypoint.sh"]
 CMD ["raindrops"]
