+++ go.mod
@@ -1,12 +1,11 @@
-module cloudmonitor
+module github.com/BouncyLlama/raindrops
 
-go 1.12
+go 1.13
 
 require (
 	github.com/PuerkitoBio/goquery v1.5.0
 	github.com/akamensky/argparse v0.0.0-20191006154803-1427fe674291
-	github.com/influxdata/influxdb-client-go v0.1.3
 	github.com/influxdata/influxdb1-client v0.0.0-20190809212627-fc22c7df067e
-	github.com/stretchr/testify v1.3.0
-	golang.org/x/net v0.0.0-20191002035440-2ec189313ef0 // indirect
+	github.com/stretchr/testify v1.4.0
+
 )
