+++ internal/testdata/azure_today_bad.html
