+++ internal/testdata/azure_today_good.html
