+++ internal/types/consts.go
@@ -1,0 +1,8 @@
+package types
+
+const Up = "up"
+const Down = "down"
+const All = "all"
+const Google = "google"
+const Amazon = "amazon"
+const Azure = "azure"
